Add files via upload
Objective: To Predict if a patient can have sepsis in advance.

But in order to do that, it is also crucial to choose appropriate variables and design valuable algorithms in the clinical setting.

The input variables of the model are physiological indicators and the output variable is whether the patient would suffer from sepsis several hours later.

Specifically, the input variables generally include vital signs like heart rate, oxygen saturation, and body temperature; biomarkers like procalcitonin and interleukin-6; laboratory values like bicarbonate and creatinine; and demographic variables like sex and age.

In most cases, the variables include lots of missing values, such as that in MIMIC III (Intensive Care Medical Information Market Database), which has been used in many studies. Among most researches, variables with lots of missing values are excluded from predictors, so valuable information may be lost as a result.

Several studies use imputation and mean filling methods to fill in missing values, but this may also lead to selection bias or mixtures of confounding factors. The data preprocessing method needs to be considered according to the characteristics of different data sets.